fix: cargo fmt on pending Rust changes
Formats all working tree changes that were pending before v0.7.0.
Fixes CI format check failure.

+++ apps/dev-cli/src/ui/event.rs
@@ -133,7 +133,7 @@ pub enum UiEvent {
         step_name: String,
         success: bool,
         duration_ms: u64,
-        tokens_used: u32,
+        tokens_used: u64,
     },
 
     /// Live progress bar during plan execution.
